requirements_render.txt
# Core web framework
Flask==2.3.3
gunicorn==21.2.0

# LangChain essentials
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.63
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.8

# Vector DB
pinecone==7.3.0

# Embeddings
sentence-transformers==2.6.0
transformers==4.57.1

# Lightweight PyTorch (CPU only - saves ~2GB memory)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.9.1+cpu

# Utilities
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.4
numpy==1.26.4
safetensors==0.7.0
tokenizers==0.22.1
huggingface-hub==0.36.0
filelock==3.20.0
fsspec==2025.10.0
sympy==1.14.0
networkx==3.4.2
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
typing-extensions==4.15.0
regex==2025.11.3
tqdm==4.67.1
scikit-learn==1.7.2
scipy==1.15.3
joblib==1.5.2
threadpoolctl==3.6.0

aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.121.3
ffmpy==1.0.0
filelock==3.20.0
Flask==2.3.3
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
gradio==3.43.0
gradio_client==0.5.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.63
langchain-huggingface==0.2.0
langchain-openai==0.3.11
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.8
langgraph-checkpoint==3.0.1
langgraph-sdk==0.2.9
langsmith==0.1.147
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pillow==10.4.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.12.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==3.17.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.6.0
sentencepiece==0.2.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==11.0.3
Werkzeug==3.1.3
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
streamlit==1.28.0